refactor(cloud): extract PostgresSessionsRepository from the store
postgres-control-plane-store.ts was a 2,718-line monolith — the single
largest modularity exception in the package. It already delegated ~21
domain families to Postgres*Repository sub-stores; the session core was
the last big inline block.

Extract the session engine (session CRUD/listing, session + workspace
event append/replay, projections, worker leases, and the command queue)
plus its private helpers into postgres-store-domains/sessions.ts as
PostgresSessionsRepository, following the existing thread-index/workflows
pattern. Tenant checks, the transaction runner, the SSE NOTIFY emitter,
the audit-event recorder and the quota deps are injected via the host;
requireSession/requireSessions/incrementSessionCounter/assertLeaseToken-
IfPresent are exposed for the store's remaining callers (channel-
interaction resolution, thread-index and workflow repo injections).

The store drops to 1,861 lines — below the 2,000 large-file threshold,
so it is no longer a documented modularity exception at all. Behaviour
is preserved: the in-memory + pglite (real Postgres SQL) control-plane
parity, concurrency, and projection suites all pass unchanged.
